Pipeline: split out ipc3 and ipc4 code from pipeline source
Split out ipc3/ipc4 code from pipeline source.
