+++ src/audio/pipeline/CMakeLists.txt
@@ -7,3 +7,8 @@ add_local_sources(sof
 	pipeline-xrun.c
 	pipeline-schedule.c
 )
+if(CONFIG_IPC_MAJOR_3)
+	add_local_sources(sof pipeline-ipc3.c)
+elseif(CONFIG_IPC_MAJOR_4)
+	add_local_sources(sof pipeline-ipc4.c)
+endif()
