+++ src/audio/pipeline/pipeline-graph.c
@@ -409,82 +409,3 @@ struct comp_dev *pipeline_get_dai_comp(uint32_t pipeline_id, int dir)
 
 	return NULL;
 }
-
-#if CONFIG_IPC_MAJOR_4
-/* Playback only: visit connected pipeline to find the dai comp and latency.
- * This function walks down through a pipelines chain looking for the target dai component.
- * Calculates the delay of each pipeline by determining the number of buffered blocks.
- */
-struct comp_dev *pipeline_get_dai_comp_latency(uint32_t pipeline_id, uint32_t *latency)
-{
-	struct ipc_comp_dev *ipc_sink;
-	struct ipc_comp_dev *ipc_source;
-	struct comp_dev *source;
-	struct ipc *ipc = ipc_get();
-
-	*latency = 0;
-
-	/* Walks through the ipc component list and get source endpoint component of the given
-	 * pipeline.
-	 */
-	ipc_source = ipc_get_ppl_src_comp(ipc, pipeline_id);
-	if (!ipc_source)
-		return NULL;
-	source = ipc_source->cd;
-
-	/* Walks through the ipc component list and get sink endpoint component of the given
-	 * pipeline. This function returns the first sink. We assume that dai is connected to pin 0.
-	 */
-	ipc_sink = ipc_get_ppl_sink_comp(ipc, pipeline_id);
-	while (ipc_sink) {
-		struct comp_buffer *buffer;
-		uint64_t input_data, output_data;
-		struct ipc4_base_module_cfg input_base_cfg = {.ibs = 0};
-		struct ipc4_base_module_cfg output_base_cfg = {.obs = 0};
-		int ret;
-
-		/* Calculate pipeline latency */
-		input_data = comp_get_total_data_processed(source, 0, true);
-		output_data = comp_get_total_data_processed(ipc_sink->cd, 0, false);
-
-		ret = comp_get_attribute(source, COMP_ATTR_BASE_CONFIG, &input_base_cfg);
-		if (ret < 0)
-			return NULL;
-
-		ret = comp_get_attribute(ipc_sink->cd, COMP_ATTR_BASE_CONFIG, &output_base_cfg);
-		if (ret < 0)
-			return NULL;
-
-		if (input_data && output_data && input_base_cfg.ibs && output_base_cfg.obs)
-			*latency += input_data / input_base_cfg.ibs -
-				output_data / output_base_cfg.obs;
-
-		/* If the component doesn't have a sink buffer, it can be a dai. */
-		if (list_is_empty(&ipc_sink->cd->bsink_list))
-			return dev_comp_type(ipc_sink->cd) == SOF_COMP_DAI ? ipc_sink->cd : NULL;
-
-		/* Get a component connected to our sink buffer - hop to a next pipeline */
-		buffer = buffer_from_list(comp_buffer_list(ipc_sink->cd, PPL_DIR_DOWNSTREAM)->next,
-					  PPL_DIR_DOWNSTREAM);
-		source = buffer_get_comp(buffer, PPL_DIR_DOWNSTREAM);
-
-		/* buffer_comp is in another pipeline and it is not complete */
-		if (!source || !source->pipeline)
-			return NULL;
-
-		/* As pipeline data is allocated in cached space, continue calculation for next
-		 * connected pipeline only if that pipeline is on same core.
-		 * This is a workaround, the real solution would be to use something like
-		 * process_on_core() to continue calculation on required core. However, as this
-		 * "latency feature" seems never used anyway, this workaround could be enough.
-		 */
-		if (!cpu_is_me(source->ipc_config.core))
-			return NULL;
-
-		/* Get a next sink component */
-		ipc_sink = ipc_get_ppl_sink_comp(ipc, source->pipeline->pipeline_id);
-	}
-
-	return NULL;
-}
-#endif
