+++ src/audio/pipeline/pipeline-stream.c
@@ -28,62 +28,6 @@
 
 LOG_MODULE_DECLARE(pipe, CONFIG_SOF_LOG_LEVEL);
 
-/*
- * Check whether pipeline is incapable of acquiring data for capture.
- *
- * If capture START/RELEASE trigger originated on dailess pipeline and reached
- * inactive pipeline as it's source, then we indicate that it's blocked.
- *
- * @param rsrc - component from remote pipeline serving as source to relevant
- *		 pipeline
- * @param ctx - trigger walk context
- * @param dir - trigger direction
- */
-static inline bool
-pipeline_should_report_enodata_on_trigger(struct comp_dev *rsrc,
-					  struct pipeline_walk_context *ctx,
-					  int dir)
-{
-	struct pipeline_data *ppl_data = ctx->comp_data;
-	struct comp_dev *pipe_source = ppl_data->start->pipeline->source_comp;
-
-	/* In IPC3, FW propagates triggers to connected pipelines, so
-	 * it can have determistic logic to conclude no data is
-	 * available.
-	 * In IPC4, host controls state of each pipeline separately,
-	 * so FW cannot reliably detect case of no data based on
-	 * observing state of src->pipeline here.
-	 */
-#if CONFIG_IPC_MAJOR_4
-	return false;
-#endif
-
-	/* only applies to capture pipelines */
-	if (dir != SOF_IPC_STREAM_CAPTURE)
-		return false;
-
-	/* only applicable on trigger start/release */
-	if (ppl_data->cmd != COMP_TRIGGER_START &&
-	    ppl_data->cmd != COMP_TRIGGER_RELEASE)
-		return false;
-
-	/* only applies for dailess pipelines */
-	if (pipe_source && dev_comp_type(pipe_source) == SOF_COMP_DAI)
-		return false;
-
-	/* source pipeline may not be active since priority is not higher than current one */
-	if (rsrc->pipeline->priority <= ppl_data->start->pipeline->priority)
-		return false;
-
-	/* if component on which we depend to provide data is inactive, then the
-	 * pipeline has no means of providing data
-	 */
-	if (rsrc->state != COMP_STATE_ACTIVE)
-		return true;
-
-	return false;
-}
-
 static int pipeline_comp_copy(struct comp_dev *current,
 			      struct comp_buffer *calling_buf,
 			      struct pipeline_walk_context *ctx, int dir)
@@ -372,16 +316,8 @@ static int pipeline_comp_trigger(struct comp_dev *current,
 			 * don't use more than one DAI per copier
 			 */
 			struct dai_data *dd;
-#if CONFIG_IPC_MAJOR_3
-			dd = comp_get_drvdata(current);
-#elif CONFIG_IPC_MAJOR_4
-			struct processing_module *mod = comp_get_drvdata(current);
-			struct copier_data *cd = module_get_private_data(mod);
 
-			dd = cd->dd[0];
-#else
-#error Unknown IPC major version
-#endif
+			dd = get_pipeline_dai_device_data(current);
 			ppl_data->delay_ms = dai_get_init_delay_ms(dd->dai);
 		}
 		break;
@@ -414,6 +350,10 @@ static int pipeline_comp_trigger(struct comp_dev *current,
 
 		return 0;
 	}
+
+	err = pipeline_is_single_triggered(is_single_ppl, is_same_sched, current, ctx, dir);
+	if (err < 0)
+		return err;
 
 	current->pipeline->trigger.pending = false;
